.golangci.yml: surpressed internal package and underscore linters
There is no rules but opinions about the use of internal package and
underscore in naming. As far as deployment concerns, both of these
are functionally related especially when it comes to constant naming.

Due to the noises created by golangci-lint and forcing everyone to
use //nolint: everywhere, it's best to surpress these false positive
opinionated linters at the root configurations level. Hence, let's
do it.

This patch surpresses internal package and underscore linters in
.golangci.yml.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>