pkg/monteur: added placeholder operation codes into repository
Instead of operating all the public API directly in the API.go and
create a very long source code that is hard to maintain, we should
create all the placeholder operation codes for it. That way, it's
easier to operate and solidify the public API calls. Hence, let's
do it.

This patch adds placeholder operation codes int pkg/monteur/
directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>