pkg/monteur/internal/libpublish: added implementation for TOMLPublisher
There is a need to split TOMLPublisher away from TOMLBuilder in order
to make things work. Otherwise, the parsed data will be ultimately
confused and hard to process. In the mean time, the implementation
for parsing TOMLPublisher data effectively is required. Hence, we
need to add the implementation for TOMLPublisher.

This patch adds implementation for TOMLPublisher in
pkg/monteur/internal/libpublish package.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>