pkg/monteur: added dep and cmd compute condition checking
The existing dependency and command's compute system condition are
unchecked. This is fatal as not all commands or dependencies are
generic across various systems. Hence, we need to add the checking
mechanism into it. Let's do it.

This patch adds dep and cmd compute conditions checking into
pkg/monteur package.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>