pkg/monteur/internal/libpublish: implemented command executions
Since we're ready to implement the list of of commands, let's
proceed to add placeholder to commander's action and implements
the commands list executions. That way, we can close the case for
publisher and proceed straight to develop all the command's algorithm.
Hence, let's do it.

This patch implements commands list executions in
pkg/monteur/internal/libpublish package.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>