docs/en/_index.md: updated landing page to secure the knowledge
Eversince the prototyping from setup to publish, I had been storing
all the knowledge in my brain to the point of forgetting. Hence,
let's update the documentations and store all the knowledge in it.

This patch updates landing page to secure the knowledge of Monteur
in docs/en/_index.md.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>