monteur/internal/libcmd/Manager.go: removed unused Metadata.Type
The Metadata.Type was initially setup for locking a recipe to a
specific usage. However, after documenting the Publish and Compose
APIs, it makes no sense to pursue that direction when a generalized
method is use. Hence, let's remove it.

This patch removes Metadata.Type from
monteur/internal/libcmd/Manager.go.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>