gopkg/monteur/internal: added conductor package
Since we are going to use asynchonous orchestration for every API
executions, it's better we abstract and generate all the orchestration
implementations into its dedicated package for upstream and Monteur
use. This way, it's sane enough to maintain all the APIs and upstream
capable for scaling to other projects. Hence, let's do it.

This patch adds condcutor package into gopkg/monteur/internal
directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>