.gitlab-ci.yml: optimized to faster running
The existing runner is overwhelmed by the updates and package
installation that isn't being used by the job. Hence, we need to
optimize it again to ensure the runner only install packages for
those that are being used. Let's do it.

This patch optimizes .gitlab-ci.yml to run faster by only install
dependencies only when the job is needed.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>