gopkg/go.mod: updated go module to match the latest and greatest
Since the creator Russ Cox pointed out that the current go module
is faulty where the module is supposedly include gopkg/ directory
into its pathing, we need to update it accordingly in order to let
go test works. Hence, let's update the go.mod to match the latest
setup.

This patch updates go module to match the latest and greatest in
gopkg/go.mod and corrected the module pathing error in its naming
as well.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>