gopkg/go.mod: pull the latest go-toml/v2 commit c862c34 for nested parse
The current go-toml/v2 package is not able to parse nested map and
arrays. Since Thomas Pelletier fixed the bug in commit 79e78b234
(see https://github.com/pelletier/go-toml/discussions/671) but not
released as of to-date, we need to pull the latest commit as of
today, which is c862c34 in order to have a working parser. Hence,
let's do it.

This patch pulls the latest go-toml/v2 commit c862c34 for nested
parsing feature in gopkg/go.mod.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>