gopkg/monteur/helpers.go: added nil logger guard to prevent panic
Since the _reportError can be called before the logger is up, it
is best we check for its nil status before use in order to prevent
a panic event. Hence, let's do it.

This patch adds nil logger checking and guarding to prevent unnecessary
panic in gopkg/monteur/helpers.go.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>