gopkg/monteur/internal/endec/toml: refactored error handling
The current TOML endec package's error handling is horrible when
it is being used across multiple packages. This is unacceptable.
Hence, we need to refactor it so that we can get meaningful error
message. Let's do it.

This patch refactors error handling to a meaningful return
in gopkg/monteur/internal/endec/toml.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>