gopkg/monteur: consolidated all gen. variable working to libworkspace
Since variables are expected to be parsed at general workspace
level and then later to the ci-job specific ones, we should
consolidate the work all the way up to the libworkspace. Once
the workspace is completed, the only remaining variables would
be the Task specific ones. This aligns to the documentations as
well. Hence, let's do it.

This patch consolidates all general variables working mechanism
to libworkspace internal package in gopkg/monteur packages.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>