gopkg/monteur/internal/archive/deb: add Generate functions to Data
Since the Data has to be able to generate all its content properly
for debhelper to operate, we have supply a function for it. Hence,
let's add a Generate() function to Data.

This patch adds Generate functions to Data data structure for
generating all the necessary files for debhelper to compile .deb
packages.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>