.configs/monteur/clean/cleaners: use delete-recursive-quiet
In case where directories are already missing and delete-recursive
is making error reporting becuase of it, we want it to be quiet.
Hence, let's refactor them to use quiet removal.

This patch refactors all recipes in .configs/monteur/clean/cleaners/
directory to use delete-recursive-quiet to slient the useless error
reporting due to the already missing directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>