gopkg/monteur/internal: fixed Copy(...) bug for windows
Windows binaries are not able to be built mainly because of the
many platform-specific time functions either from file or symlink.
Hence, we need to fix them in order to enable windows-amd64 and
windows-arm64 build. Let's do it.

This patch fixed Copy(...) function bug for building windows images
in gopkg/monteur/internal/* packages.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>