docs/en/ci-jobs: added 'Prepare' CI Job documentation
Since there are jobs that are not suitable to be with build and
package CI Jobs like changelog generations, we should always
split them away to make things crisp and clear. Hence, let's document
one.

This patch adds 'Prepare' CI Job documentation into the repository
as the primary design guidelines.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>