gopkg: added Prepare API
Since we need a dedicated API for tasks that are not suitable to
be inside Build API or Package API, we have to facilitate one for
it. Since both Build API and Package API needs it, we call this as
Prepare API. Hence, let's do it.

This patch adds Prepare API into gopkg/ packages.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>