docs/en: updated landing page with latest migrations and format
The existing landing page is still using the rendering prior to
GitHub migrations. Hence, we should update it to the latest
migrations and format. Let's do this.

This patch updates docs/en landing page _index.html file with the
latest migration and its HTML format over the previous Markdown
format.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>