docs/.layouts: added epilogue shortcode
Since every Epilogue directs visitors to external channels like
GitLab Issues section, to avoid vendor lock-in, it's better we
abstract them out and create a shortcode for it. That way, in case
we change vendor, it's easier for us to roll-out the changes. Hence,
let's do this.

This patch adds epilogue shortcode in docs/.layouts/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>