docs/en/internals/secrets: updated to use latest rendering
The existing secrets page is out of date and is still using the
old rendering. Moreover, some content were written wrongly. Hence,
we need to roll out a new update in conjunction with the migrations.
Let's do this.

This patch updates docs/en/internals/secrets/ directory to use the
latest rendering.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>