docs/en/internals/package-meta: updated to use latest rendering
The existing package-meta page is outdated, contains a lot of
content errors all over the places, and is using the older logo
in its thumbnails before the migrations. Hence, we need to update
it. Let's do this.

This patch updates docs/en/internals/package-meta to use the latest
rendering after migrations.

Signed-off-by: (Holloway) Chew, Kean Ho <kean.ho.chew@zoralab.com>